Move Command to Framework
